Cadastro de Despesas

+++ app/Http/Controllers/Expense/ExpenseController.php
@@ -4,12 +4,18 @@ namespace App\Http\Controllers\Expense;
 
 use App\Http\Controllers\Controller;
 use App\Models\Expense;
+use App\Models\ExpenseClient;
+use App\Models\ExpenseUser;
 use App\Models\User;
 use App\Models\Category;
 use App\Models\PaymentMethod;
 use App\Models\Client;
+use App\Models\Authorization;
 use App\Http\Requests\ExpenseRequest;
 use App\Helpers\AuthorizationHelper;
+use App\Helpers\ExpenseHelper;
+use App\Helpers\DateTimeHelper;
+use Illuminate\Support\Facades\Auth;
 use Carbon\Carbon;
 use DataTables;
 
@@ -47,8 +53,69 @@ class ExpenseController extends Controller
             return redirect()->back()->with('error', 'Você não tem permissão para cadastrar nessa página!')->withInput();
         }
 
+        $count = Authorization::where('id_authorization', $request->input('id_authorization'))
+            ->where('active', true)
+            ->where('approved', true)
+            ->count();
+        if ($count <= 0) {
+            return redirect()->back()->with('error', 'Autorização não encontrada. Verifique se a mesma existe, e se está ativa e aprovada!')->withInput();
+        }
+
+        $authorization = Authorization::where('id_authorization', $request->input('id_authorization'))->first();
+
         try {
-            $expense = Expense::create($request->all());
+            $dates = DateTimeHelper::distribute($request->amount, $request->date, $request->distribute, $authorization->end_date, Auth::user()->id_branch);
+
+            if ($request->distribute <= 1 || count($dates) <= 0) {
+                $expense = Expense::create($request->all());
+                $this->expensesClients($expense->id_expense, $request->client_amount);
+                $this->expensesUsers($expense->id_expense, $request->user_amount);
+                ExpenseHelper::refresh($expense->id_expense);
+            } else {
+
+                foreach ($dates as $date => $amount) {
+                    $request['date'] = $date;
+                    $request['amount'] = $amount;
+                    $expense = Expense::create($request->all());
+
+                    $accumulated_amount = 0;
+                    $client_amount = $request->client_amount;
+                    $updated_client_amount = [];
+                    $last_id_client = 0;
+                    foreach ($client_amount as $id_client => $partial_amount) {
+                        if ($partial_amount) {
+                            $partial_amount = round($amount * $request->client_percentage[$id_client] / 100, 2);
+                            $accumulated_amount += $partial_amount;
+                            $updated_client_amount[$id_client] = $partial_amount;
+                            $last_id_client = $id_client;
+                        }
+                    }
+                    $updated_client_amount[$last_id_client] += round($amount - $accumulated_amount, 2);
+                    $request->merge(['client_amount' => $updated_client_amount]);
+
+                    $this->expensesClients($expense->id_expense, $request->client_amount);
+
+
+                    $accumulated_amount = 0;
+                    $user_amount = $request->user_amount;
+                    $updated_user_amount = [];
+                    $last_id_user = 0;
+                    foreach ($user_amount as $id_user => $partial_amount) {
+                        if ($partial_amount) {
+                            $partial_amount = round($amount * $request->user_percentage[$id_user] / 100, 2);
+                            $accumulated_amount += $partial_amount;
+                            $updated_user_amount[$id_user] = $partial_amount;
+                            $last_id_user = $id_user;
+                        }
+                    }
+                    $updated_user_amount[$last_id_user] += round($amount - $accumulated_amount, 2);
+                    $request->merge(['user_amount' => $updated_user_amount]);
+
+                    $this->expensesUsers($expense->id_expense, $request->user_amount);
+                    ExpenseHelper::refresh($expense->id_expense);
+                }
+            }
+
             return redirect()->route('expenses.show', ['id' => $expense->id_expense])->with('success', 'Registro cadastrado com sucesso');
         } catch (\Exception $e) {
             return redirect()->back()->with('error', $e->getMessage())->withInput();
@@ -63,8 +130,15 @@ class ExpenseController extends Controller
      */
     public function show($id)
     {
-        $data = Expense::with('authorization')->find($id);
+        $data = Expense::with(['authorization', 'clients', 'users'])
+            ->where(['id_expense' => $id, 'id_user' => Auth::id()])->whereNull('id_batch')->first();
         if ($data) {
+            $clientsById = array_column($data->clients->toArray(), null, 'id_client');
+            $data->clients = $clientsById;
+
+            $usersById = array_column($data->users->toArray(), null, 'id_user');
+            $data->users = $usersById;
+
             $authorizations = AuthorizationHelper::active('expense');
             $categories = Category::orderBy('name')->with('category_type')
                 ->whereHas('category_type', function ($query) {
@@ -93,17 +167,17 @@ class ExpenseController extends Controller
             return redirect()->back()->with('error', 'Você não tem permissão para salvar nessa página!')->withInput();
         }
 
-        if ($request->_action == "store") {
-            $id = null;
-            $storeRequest = new ExpenseRequest();
-            $request->validate($storeRequest->rules());
-            $storeRequest->merge($request->all());
-            return $this->store($storeRequest);
-        }
+        unset($request['id_authorization']);
+
         try {
-            $expense = Expense::find($id);
+            $expense = Expense::where(['id_expense' => $id, 'id_user' => Auth::id()])->whereNull('id_batch')->first();
             if ($expense) {
+
                 $expense->update($request->all());
+                $this->expensesClients($expense->id_expense, $request->client_amount);
+                $this->expensesUsers($expense->id_expense, $request->user_amount);
+                ExpenseHelper::refresh($expense->id_expense);
+
                 return redirect()->route('expenses.show', ['id' => $expense->id_expense])->with('success', 'Registro salvo com sucesso');
             } else {
                 return redirect()->route('expenses')->with('error', 'Registro não encontrado!');
@@ -126,7 +200,11 @@ class ExpenseController extends Controller
         }
 
         try {
-            $expense = Expense::find($id);
+            $expense = Expense::where(['id_expense' => $id, 'id_user' => Auth::id()])->whereNull('id_batch')->first();
+            if ($expense->id_batch) {
+                return redirect()->back()->with('error', 'Esta despesa está vinculada a um lote. Não é possível apagá-la!')->withInput();
+            }
+
             if ($expense) {
                 $expense->delete();
                 return redirect()->route('expenses')->with('success', 'Registro apagado com sucesso');
@@ -141,7 +219,7 @@ class ExpenseController extends Controller
 
     public function datatable()
     {
-        $data = Expense::with('category')->latest()->get();
+        $data = Expense::with('category')->where(['id_user' => Auth::id()])->whereNull('id_batch')->latest()->get();
         $id_field = request('id-field') ?: 'id';
 
         return DataTables::of($data)
@@ -165,5 +243,40 @@ class ExpenseController extends Controller
             })
             ->rawColumns(['actions', 'date'])
             ->make(true);
+    }
+
+    public function expensesClients($id_expense, $clients)
+    {
+        $expense = Expense::where('id_expense', $id_expense)->first();
+        $authorization = Authorization::where('id_authorization', $expense->id_authorization)
+            ->with('clients')->first();
+
+        $clients_list = $authorization->clients->pluck('id_client')->toArray();
+        $clients = array_intersect_key($clients, array_flip($clients_list));
+
+        ExpenseClient::where('id_expense', $id_expense)->delete();
+        if ($clients && count($clients) > 0) {
+            foreach ($clients as $id_client => $amount) {
+                if ($amount > 0) {
+                    $percentage = $amount / $expense->amount * 100;
+                    ExpenseClient::create(['id_expense' => $id_expense, 'id_client' => $id_client, 'amount' => $amount, 'percentage' => $percentage]);
+                }
+            }
+        }
+    }
+
+    public function expensesUsers($id_expense, $users)
+    {
+        $expense = Expense::where('id_expense', $id_expense)->first();
+
+        ExpenseUser::where('id_expense', $id_expense)->delete();
+        if ($users && count($users) > 0) {
+            foreach ($users as $id_user => $amount) {
+                if ($amount > 0) {
+                    $percentage = $amount / $expense->amount * 100;
+                    ExpenseUser::create(['id_expense' => $id_expense, 'id_user' => $id_user, 'amount' => $amount, 'percentage' => $percentage]);
+                }
+            }
+        }
     }
 }
